serialqueue: add some logging to identify MCU disconnect
Something funny is going on with serial transfer.   The logs show the following just before a MCU timeout:

Got error 0 in read: (0)Success

I suspect the source is serialqueue.c, line 543.  The read is returning 0, which should signify a closed serial port.  Most likely something on the MCU side is closing the port.