a4954: implement move_to_phase command
The current "set_phase" command relies on the previous phase to determine how to set the gpios.  Unfortunately this does not work when stepping via closed loop, as the previously recorded phase is not likely to match the stepper position,

The new implementation is very similar to set_phase.  The biggest difference is that it tracks the state of the gpios,  so knowledge of the previous position isn't required to toggle them.