mechaduino:  Add SERVO_ENABLE gcode
When in PID mode the servo stepper needs initialization before it can process incoming step commands.    When allowing Klipper to automatic transition from disabled to enabled this causes an initial jerk as the stepper processes the steps already in the step queue.

This is a workaround, ultimately it would be better if functionality was added to klipper that adds some delay to the scheduling of steps when using a servo stepper.